+++ NEWS.md
@@ -1,5 +1,7 @@
 # rtweet (development version)
 
+* Fix `auth_sitrep()` to work well with OAuth2 tokens.
+
 # rtweet 1.2.0
 
 ## Authentication changes